requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.23.5
pandas==1.5.3
matplotlib==3.6.3
tqdm==4.64.1
scikit-learn==1.2.1
seaborn==0.12.2
Pillow==9.4.0
